pyproject.toml
[project]
name = "health_tracker"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "nazar@foxcom.lv", email = "nazar@foxcom.lv"},
]
dependencies = [
    "fastapi[all]>=0.113.0",
    "pydantic-settings>=2.4.0",
    "python-json-logger>=2.0.7",
    "sqlalchemy>=2.0.34",
    "asyncpg>=0.29.0",
    "alembic>=1.13.2",
    "psycopg2-binary>=2.9.9",
    "psycopg2>=2.9.9",
    "aiohttp>=3.10.5",
    "phonenumbers>=8.13.45",
]
requires-python = "==3.12.*"
readme = "README.md"
license = {text = "MIT"}


[tool.pdm]
distribution = false
